Update README.md
Q1. Open the data BraceroData.dta that I posted on Canvas (note: BraceroData_s13.dta can be opened in Stata 13 and is posted under the [data] folder). This data is from the paper, “Immigration Restrictions as Active Labor Market Policy: Evidence from the Mexican Bracero Exclusion” (posted on Canvas as BraceroPaper.pdf) and modified slightly for ease-of-use with this problem set.

The bracero program was a series of bilateral agreements between the US and Mexico that permitted temporary immigration of agricultural workers from Mexico to the US in the 1940s-1960s. At its peak, these agreements were responsible for the supply of half a million seasonal workers each year to US farms, with work contracts for six weeks to six months. This program was restricted in 1962 and fully terminated on December 31, 1964 with the stated goal of “improving wages and domestic employment for seasonal farm workers in the US.” In other words, the people who advocated for terminating this program did so because they believed that by restricting the immigration of Mexican workers in this sector, the farms in the US would increase their demand for US workers and increase wages for those workers.

However, because farms can substitute between labor and capital, it is not clear what the effect of this policy will be. When farms can no longer employ immigrant workers through the bracero program, will they employ more domestic workers (and increase wages) or will they increase their use of technology so that the same work can be done with fewer workers? Using this data and a difference-in-differences research design, you can test whether ending the bracero program increased employment of domestic seasonal workers and increased wages. In the data, I have already defined treated = 1 as states that were heavily reliant on bracero workers prior to the ending of the program and treated = 0 as states that were not heavily reliant on bracero workers prior to the ending of the program.